newDesignActivity

+++ app/src/main/java/com/example/yazon/weather/responseWeather/Main.java
@@ -6,8 +6,8 @@ import java.io.Serializable;
 
 public class Main implements Serializable {
     public Float temp;
-    public Double pressure;
-    public Double humidity;
+    public Float pressure;
+    public Float humidity;
     public Double temp_min;
     public Double temp_max;
